backend/requirements.txt
fastapi
uvicorn[standard]
openai
youtube-transcript-api
python-dotenv